refactor: Restaurant ResponseDTO 패키지 이동 (#225)
